requirements.txt
python-telegram-bot==20.3
pymongo[srv]==4.3.3
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
dnspython==2.4.2